Refactor: [BADA-119] features/trade 폴더 fsd 아키텍처 구조로 변경 #53
