+++ src/features/trade/api/trade.ts
@@ -1,6 +1,6 @@
 import { axiosInstance } from '@lib/axios/axiosInstance';
 import { END_POINTS } from '@constants/api';
-import { Post, TradeDetailPost } from '@features/trade/models/post';
+import { Post, TradeDetailPost } from '@features/trade/lib/types';
 
 export const getTradePosts = async (): Promise<Post[]> => {
   const content: { postsResponse: Post[] } = await axiosInstance.get(END_POINTS.TRADES.LIST);
