+++ src/features/trade/ui/TradeCouponList.tsx
@@ -1,41 +1,46 @@
 import { ArrowDownUp, ListFilter } from 'lucide-react';
 import { Product } from '@ui/Product';
 import { Drawer, DrawerButton, FilterDrawerButton } from '@ui/Drawer';
-import { useTradePostsQuery } from '@features/trade/queries/useTradeQuery';
+import { useTradePostsQuery } from '@features/trade/model/useTradeQueries';
 
-type SortOption = 'latest' | 'popular';
-
-interface TradeDataListProps {
-  sortOption: SortOption;
-  setSortOption: (value: SortOption) => void;
+interface TradeCouponListProps {
+  sortOption: 'latest' | 'popular';
+  setSortOption: (value: 'latest' | 'popular') => void;
   isSortDrawerOpen: boolean;
   setIsSortDrawerOpen: (value: boolean) => void;
+  selectedCategory: string;
 }
 
-export function TradeDataList({
+export function TradeCouponList({
   sortOption,
   setSortOption,
   isSortDrawerOpen,
   setIsSortDrawerOpen,
-}: TradeDataListProps) {
+  selectedCategory,
+}: TradeCouponListProps) {
   const { posts, isLoading } = useTradePostsQuery();
 
   if (isLoading) {
-    return <div className="py-4 text-center text-[var(--black)]">로딩 중...</div>;
-  }
-  if (!posts || posts.length === 0) {
-    return <div className="py-4 text-center text-[var(--black)]">게시물이 없습니다.</div>;
+    return <div>로딩 중...</div>;
   }
 
-  const dataPosts = posts.filter((p) => p.postCategory === 'DATA');
-  const sorted = [...dataPosts].sort((a, b) =>
+  if (!posts || posts.length === 0) {
+    return <div>쿠폰 게시물이 없습니다.</div>;
+  }
+  const couponPosts = posts.filter((p) => p.postCategory === 'GIFTICON');
+
+  const filtered = couponPosts.filter(
+    (p) => selectedCategory === '전체' || p.gifticonCategory === selectedCategory,
+  );
+
+  const sorted = [...filtered].sort((a, b) =>
     sortOption === 'latest'
       ? new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
       : b.likesCount - a.likesCount,
   );
 
   return (
-    <section className="bg-white px-6">
+    <section className="bg-white px-6 py-4">
       <div className="flex flex-row justify-between py-2">
         <button
           onClick={() => setIsSortDrawerOpen(true)}
@@ -50,10 +55,11 @@ export function TradeDataList({
           <ListFilter size={14} />
         </div>
       </div>
-      <div className="flex flex-col gap-4">
+      <div className="flex flex-col gap-4 py-4">
         {sorted.map((item) => (
           <Product
             key={item.id}
+            brand={item.partner}
             name={item.title}
             price={item.price}
             imageSrc={item.postImage}
@@ -61,7 +67,6 @@ export function TradeDataList({
           />
         ))}
       </div>
-
       <Drawer isOpen={isSortDrawerOpen} onClose={() => setIsSortDrawerOpen(false)} variant="filter">
         <FilterDrawerButton
           selected={sortOption === 'latest'}
