+++ src/features/trade/ui/detail/TradeDetailDrawer.tsx
