+++ src/features/trade/ui/detail/TradeDetailProductSection.tsx
