+++ src/features/trade/ui/detail/TradeDetailSellerSection.tsx
@@ -4,7 +4,7 @@ import { ImageBox } from '@ui/ImageBox';
 import { LikeButton } from '@ui/LikeButton';
 import { ProductInfo } from '@ui/ProductInfo';
 import { Profile } from '@ui/Profile';
-import { useUserTradePostsQuery } from '@features/trade/queries/useUserTradePostQuery';
+import { useUserTradePostsQuery } from '@features/trade/model/useUserTradePostQuery';
 
 interface TradeDetailSellerSectionProps {
   sellerId: number; // 판매자 userId
