+++ src/features/trade/ui/search/SearchHeader.tsx
