+++ src/features/trade/ui/search/SearchHotKeywords.tsx
