+++ src/features/trade/ui/search/SearchRecentKeywords.tsx
