+++ src/features/trade/ui/search/SearchResult.tsx
